Revert preset renaming
