+++ src/generators/java/presets/index.ts
@@ -1,4 +1,4 @@
 export * from './CommonPreset';
 export * from './DescriptionPreset';
-export * from './JavaJacksonPreset';
-export * from './JavaConstraintsPreset';
+export * from './JacksonPreset';
+export * from './ConstraintsPreset';
