+++ test/generators/java/presets/ConstraintsPreset.spec.ts
